+++ demos/kdf/argon2.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 The OpenSSL Project Authors. All Rights Reserved.
+ * Copyright 2023-2025 The OpenSSL Project Authors. All Rights Reserved.
  *
  * Licensed under the Apache License 2.0 (the "License").  You may not use
  * this file except in compliance with the License.  You can obtain a copy
@@ -113,7 +113,7 @@ int main(int argc, char **argv)
         if (max_threads == 0)
             threads = 1;
         else if (max_threads < parallel_cost)
-            threads = max_threads;
+            threads = (unsigned int)max_threads;
     }
 
     /* Set password */
