+++ demos/keyexch/x25519.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022-2023 The OpenSSL Project Authors. All Rights Reserved.
+ * Copyright 2022-2025 The OpenSSL Project Authors. All Rights Reserved.
  *
  * Licensed under the Apache License 2.0 (the "License").  You may not use
  * this file except in compliance with the License.  You can obtain a copy
@@ -185,7 +185,7 @@ static int keyexch_x25519_after(
     }
 
     printf("Shared secret (%s):\n", local_peer->name);
-    BIO_dump_indent_fp(stdout, local_peer->secret, local_peer->secret_len, 2);
+    BIO_dump_indent_fp(stdout, local_peer->secret, (int)local_peer->secret_len, 2);
     putchar('\n');
 
     ret = 1;
