+++ demos/mac/poly1305.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021-2023 The OpenSSL Project Authors. All Rights Reserved.
+ * Copyright 2021-2025 The OpenSSL Project Authors. All Rights Reserved.
  *
  * Licensed under the Apache License 2.0 (the "License").  You may not use
  * this file except in compliance with the License.  You can obtain a copy
@@ -183,7 +183,7 @@ int main(int argc, char **argv)
     }
 
     printf("Generated MAC:\n");
-    BIO_dump_indent_fp(stdout, out, out_len, 2);
+    BIO_dump_indent_fp(stdout, out, (int)out_len, 2);
     putchar('\n');
 
     if (out_len != sizeof(expected_output)) {
