+++ demos/signature/EVP_EC_Signature_demo.c
@@ -1,5 +1,5 @@
 /*-
- * Copyright 2021-2023 The OpenSSL Project Authors. All Rights Reserved.
+ * Copyright 2021-2025 The OpenSSL Project Authors. All Rights Reserved.
  *
  * Licensed under the Apache License 2.0 (the "License").  You may not use
  * this file except in compliance with the License.  You can obtain a copy
@@ -134,7 +134,7 @@ static int demo_sign(OSSL_LIB_CTX *libctx,  const char *sig_name,
     *sig_out_len = sig_len;
     *sig_out_value = sig_value;
     fprintf(stdout, "Generating signature:\n");
-    BIO_dump_indent_fp(stdout, sig_value, sig_len, 2);
+    BIO_dump_indent_fp(stdout, sig_value, (int)sig_len, 2);
     fprintf(stdout, "\n");
     ret = 1;
