+++ test/bio_core_test.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021-2023 The OpenSSL Project Authors. All Rights Reserved.
+ * Copyright 2021-2025 The OpenSSL Project Authors. All Rights Reserved.
  *
  * Licensed under the Apache License 2.0 (the "License").  You may not use
  * this file except in compliance with the License.  You can obtain a copy
@@ -93,7 +93,7 @@ static int test_bio_core(void)
         goto err;
 
     buf[0] = '\0';
-    if (!TEST_int_gt(BIO_write(cbio, msg, strlen(msg) + 1), 0)
+    if (!TEST_int_gt(BIO_write(cbio, msg, (int)(strlen(msg) + 1)), 0)
             || !TEST_int_gt(BIO_read(cbio, buf, sizeof(buf)), 0)
             || !TEST_str_eq(buf, msg))
         goto err;
